config/functions: After PEP-3147, make Python v3 think we have sources
https://www.python.org/dev/peps/pep-3147/#case-4-legacy-pyc-files-and-source-less-imports

Alternative is to use -b when compileall.py is executed in order to create the pyc where the
source woud have been, but not all packages support this option so we'd need to patch them.

Creating empty source files keeps the Python3 import loader happy, without wasting space (ignoring
the space taken by the filenames themselves).