Python: disable stale cache check for /usr, drop Buildroot PEP-3147 fix
The Buildroot fix seems to cause more problems than it solves. This fix
is more reliable, but messy. And only really works for LibreELEC as we know
/usr is read-only and there's no point validating the cached files.

Without this change, any cached but stale - due to mismatched timestamps and
file size - bytecode is recompiled, however as the filesystem is read-only the
new bytecode can't be written (which fails silently) so there is no usable
bytecode for Python to load/import.